requirements.txt
flask
pandas
openpyxl
flask_sqlalchemy
python-dotenv
mysql-connector-python
bcrypt